requirements.txt
streamlit>=1.28.0
pandas>=2.0.0
plotly>=6.1.1
numpy>=1.24.0
requests>=2.31.0
python-docx>=0.8.11
pdfplumber>=0.9.0
PyMuPDF>=1.23.0
python-pptx>=0.6.21
beautifulsoup4>=4.12.0
openpyxl>=3.1.0
reportlab>=4.0.0
Pillow>=10.0.0
huggingface-hub>=0.17.0
kaleido>=0.2.1